Embed default net, and simplify using non-default nets
covers the most important cases from the user perspective:

It embeds the default net in the binary, so a download of that binary
will result
in a working engine with the default net. The engine will be functional
in the default mode
without any additional user action.

It allows non-default nets to be used, which will be looked for in up to
three directories (working directory, location of the binary, and
optionally a specific default directory).
This mechanism is also kept for those developers that use MSVC,
the one compiler that doesn't have an easy mechanism for embedding data.

It is possible to disable embedding, and instead specify a specific
directory, e.g. linux distros might want to use
CXXFLAGS="-DNNUE_EMBEDDING_OFF
-DDEFAULT_NNUE_DIRECTORY=/usr/share/games/stockfish/" make -j
ARCH=x86-64 profile-build

passed STC non-regression:
https://tests.stockfishchess.org/tests/view/5f4a581c150f0aef5f8ae03a
LLR: 2.95 (-2.94,2.94) {-1.25,-0.25}
Total: 66928 W: 7202 L: 7147 D: 52579
Ptnml(0-2): 291, 5309, 22211, 5360, 293

closes #3070

fixes #3030

No functional change.
master (#3133)  sf_12
@snicolet
@vondele
snicolet authored